+++ Frontend/include/model.h
@@ -28,7 +28,7 @@ class SocketManager;
 class DataManager;
 class ISocket;
 
-using ChatId          = int;
+using ChatId          = long long;
 using ChatPtr         = std::shared_ptr<ChatBase>;
 using MessageModelPtr = std::shared_ptr<MessageModel>;
 
@@ -45,49 +45,49 @@ class Model : public QObject {
 
   ChatModel*                    getChatModel() const;
   UserModel*                    getUserModel() const;
-  MessageModel*                 getMessageModel(int chat_id);
+  MessageModel*                 getMessageModel(long long chat_id);
 
   void checkToken();
   void signIn(const LogInRequest& login_request);
   void signUp(const SignUpRequest& request);
-  void initSocket(int user_id);
+  void initSocket(long long user_id);
   void connectSocket();
 
-  [[nodiscard]] ChatPtr     loadChat(int chat_id);
+  [[nodiscard]] ChatPtr     loadChat(long long chat_id);
   [[nodiscard]] QList<User> findUsers(const QString& text);
-  std::optional<User>       getUser(int user_id);
-  ChatPtr                   createPrivateChat(int user_id);
+  std::optional<User>       getUser(long long user_id);
+  ChatPtr                   createPrivateChat(long long user_id);
 
-  QList<Message> getChatMessages(int chat_id, int limit = 20);
+  QList<Message> getChatMessages(long long chat_id, int limit = 20);
 
   void                         sendMessage(const Message& msg);
   [[nodiscard]] QList<ChatPtr> loadChats();
   void                         authenticateWithToken(const QString& token);
-  void                         getUserAsync(int user_id);
-  void                         fillChatHistory(int chat_id);
+  void                         getUserAsync(long long user_id);
+  void                         fillChatHistory(long long chat_id);
   void                         addChat(const ChatPtr& chat);
   void                         addChatInFront(const ChatPtr& chat);
-  void                         createChat(int chat_id);
-  void    addMessageToChat(int chat_id, const Message& msg);
-  void    addOfflineMessageToChat(int chat_id, User, const Message&);
+  void                         createChat(long long chat_id);
+  void    addMessageToChat(long long chat_id, const Message& msg);
+  void    addOfflineMessageToChat(long long chat_id, User, const Message&);
   void    deleteToken() const;
-  ChatPtr getPrivateChatWithUser(int user_id);
+  ChatPtr getPrivateChatWithUser(long long user_id);
   void    saveToken(const QString& token);
   void    clearAllChats();
   void    clearAllMessages();
   void    logout();
-  ChatPtr getChat(int chat_id);
+  ChatPtr getChat(long long chat_id);
 
   int         getNumberOfExistingChats() const;
-  QModelIndex indexByChatId(int chat_id);
-  static void setCurrentUserId(int current_id);
+  QModelIndex indexByChatId(long long chat_id);
+  static void setCurrentUserId(long long current_id);
 
  Q_SIGNALS:
-  void chatAdded(int id);
+  void chatAdded(long long id);
   void errorOccurred(const QString& error);
   void userCreated(const User& user, const QString& token);
   void newResponce(QJsonObject& message);
-  void chatUpdated(int chat_id);
+  void chatUpdated(long long chat_id);
 
  private:
   void onMessageReceived(const QString& msg);
